[MCUX-83782][LPSPI] Fixed MISRA C-2012 violations
- Rule 10.3 (fsl_lpspi.c:1425): add explicit (uint16_t) cast when
  assigning the TCR frame size field to the uint16_t handle member
  to resolve narrowing essential-type violation
- Rule 12.2 (fsl_lpspi.h:891): cast ternary boolean operands to
  uint32_t before shifting by LPSPI_CR_RTF_SHIFT/LPSPI_CR_RRF_SHIFT
  so the shift amount lies within [0, 31] of the left operand's
  essential type
- Rule 10.7 (fsl_lpspi.h:891): same uint32_t casts resolve the
  composite-expression essential-type width mismatch in the bitwise OR
- Rule 10.6 (fsl_lpspi_edma.c:241, 995): cast handle->fifoSize to
  uint32_t before subtracting 1U so the composite expression does not
  widen implicitly when passed to the uint32_t parameter of
  LPSPI_SetFifoWatermarks

+++ drivers/lpspi/fsl_lpspi.c
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2015, Freescale Semiconductor, Inc.
- * Copyright 2016-2022, 2024-2025 NXP
+ * Copyright 2016-2022, 2024-2026 NXP
  * All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -1422,7 +1422,7 @@ status_t LPSPI_MasterTransferNonBlocking(LPSPI_Type *base, lpspi_master_handle_t
     handle->writeTcrInIsr        = false;
 
     /* Backup frame size */
-    handle->frameSize     = (LPSPI_GetTcr(base) & LPSPI_TCR_FRAMESZ_MASK) >> LPSPI_TCR_FRAMESZ_SHIFT;
+    handle->frameSize     = (uint16_t)((LPSPI_GetTcr(base) & LPSPI_TCR_FRAMESZ_MASK) >> LPSPI_TCR_FRAMESZ_SHIFT);
     uint32_t frameSizeNew = handle->frameSize;
     handle->bytesPerFrame = (uint16_t)(handle->frameSize / 8U) + 1U;
